+++ src/components/common/Badge/Badge.type.ts
@@ -1,0 +1,4 @@
+export interface BadgeProps {
+  color: string;
+  bgColor: string;
+}
