+++ src/type/CartItem.ts
@@ -1,0 +1,7 @@
+import { ProductItem } from './ProductItem';
+
+export interface CartItems {
+  id: number;
+  quantity: number;
+  product: ProductItem;
+}
