+++ src/type/ProductItem.ts
@@ -1,0 +1,7 @@
+export interface ProductItem {
+  id: number;
+  name: string;
+  price: number;
+  imageUrl: string;
+  category: string;
+}
